f: kakao 버튼 사이즈 조절, footer 마진 조절
