+++ app/components/hidden-container/ClearBoxContainer.tsx
@@ -7,12 +7,11 @@ import useGetClipBoardText from "@/util/hooks/useGetClipBoardText";
 import { TodayIndexType } from "../main-container/page-container/MainContainer";
 import ComposedChartContainer from "./ComposedChartContainer";
 import { SignInBtn } from "../main-container/page-container/SignBtn";
-import { useRouter } from "next/navigation";
 import KakaoShare from "./KakaoShareBtn";
+import RefreshBtnContainer from "../main-container/under-main-container.tsx/RefreshBtnContainer";
 
 export default function ClearBoxContainer(props :TodayIndexType){
 
-    const router = useRouter();
     const { winState } = useWinStateLocalstorage();
     const { nowMode } = useNowMode();
     const { WIN_TEXT, LOSE_TEXT, indexGuesses, hours, minutes } = useGetClipBoardText(props.index);
@@ -24,19 +23,7 @@ export default function ClearBoxContainer(props :TodayIndexType){
             "p-3 mt-3 dark-mode-clear-box":
             "p-3 mt-3 light-mode-clear-box"
         }>
-            <button 
-                className={
-                    nowMode.mode === 'dark'? 
-                    "float-end refresh-btn border-1 rounded-1 pt-1 pb-1 dark-mode-input-and-btn":
-                    "float-end refresh-btn border-1 rounded-1 pt-1 pb-1"
-                }
-                onClick={() => {router.refresh()}}
-            >
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-arrow-clockwise" viewBox="0 0 16 16">
-                    <path fillRule="evenodd" d="M8 3a5 5 0 1 0 4.546 2.914.5.5 0 0 1 .908-.417A6 6 0 1 1 8 2z"/>
-                    <path d="M8 4.466V.534a.25.25 0 0 1 .41-.192l2.36 1.966c.12.1.12.284 0 .384L8.41 4.658A.25.25 0 0 1 8 4.466"/>
-                </svg>
-            </button>
+            <RefreshBtnContainer />
             {
                 winState?
                 <strong className="mb-2"><p className="m-0">정답 단어를 맞혔습니다.</p> {indexGuesses}번째 추측만에 정답을 맞혔네요!</strong>:
