+++ app/globals.css
@@ -300,12 +300,13 @@ body .main-container{
 }
 
 .kakao-btn{
-  width : 50px;
+  width : 60px;
+  height: auto;
   @media screen and (max-width : 470px) {
-    width : 40px;
+    width : 50px;
   }
   @media screen and (max-width : 380px) {
-    width : 30px;
+    width : 40px;
   }
 }
