+++ app/page.tsx
@@ -2,10 +2,9 @@ import Header from "./components/main-container/under-main-container.tsx/Header"
 import SettingAlertContainer from "./components/main-container/page-container/SettingAlertContainer";
 import Navbar from "./components/main-container/page-container/Navbar";
 import MainContainer from "./components/main-container/page-container/MainContainer";
-import SaveSimilarWordsContainer from "./components/SaveSimilarWordsContainer";
+import FooterContainer from "./components/main-container/page-container/Footer";
 import { cookies } from "next/headers";
 import { getServerUserData, UserDataType } from "@/util/functions/getServerUserData";
-import FooterContainer from "./components/main-container/page-container/Footer";
 
 export default async function Home() {
 
@@ -17,7 +16,6 @@ export default async function Home() {
   return (
     <>
       <Navbar />
-      {/* <SaveSimilarWordsContainer /> */}
       <SettingAlertContainer />  
       <div className="main-container">
         <Header />
